+++ changelog.d/20231006_154216_sirosen_fix_userinfo.rst
@@ -1,0 +1,5 @@
+Deprecated
+~~~~~~~~~~
+
+- ``AuthClient.oauth2_userinfo`` method has been deprecated in favor of
+  ``AuthClient.userinfo``. Callers should prefer the new method name. (:pr:`NUMBER`)
