+++ src/globus_sdk/_testing/data/auth/oauth2_userinfo.py
@@ -1,26 +1,5 @@
-from globus_sdk._testing.models import RegisteredResponse, ResponseSet
+# this is a clone of the userinfo.py data for compatibility across testing
+# it should be removed in a future release
+from .userinfo import RESPONSES
 
-from ._common import FORBIDDEN_AUTH_RESPONSE, UNAUTHORIZED_AUTH_RESPONSE
-
-RESPONSES = ResponseSet(
-    unauthorized=RegisteredResponse(
-        service="auth",
-        path="/v2/oauth2/userinfo",
-        status=401,
-        json=UNAUTHORIZED_AUTH_RESPONSE.json,
-        metadata={
-            "error_id": UNAUTHORIZED_AUTH_RESPONSE.error_id,
-            **UNAUTHORIZED_AUTH_RESPONSE.metadata_include,
-        },
-    ),
-    forbidden=RegisteredResponse(
-        service="auth",
-        path="/v2/oauth2/userinfo",
-        status=403,
-        json=FORBIDDEN_AUTH_RESPONSE.json,
-        metadata={
-            "error_id": FORBIDDEN_AUTH_RESPONSE.error_id,
-            **FORBIDDEN_AUTH_RESPONSE.metadata_include,
-        },
-    ),
-)
+__all__ = ("RESPONSES",)
